chore(warnings): kotlin compile sweep, 100 to 0
Cross-module warning cleanup. Net effect: `./gradlew :app:compileDebugKotlin`
goes from roughly 100 warnings on a clean build to zero. Three classes of fix:

Mechanical type-system cleanups (no behaviour change):
* Drop redundant `?.` safe calls on receivers the compiler has smart-cast to
  non-null: ResponseBody across the OkHttp call sites in ai/providers,
  search services, and TTS providers; UIMessagePart.Text in tool runners;
  JsonObject in CronJobScheduler-adjacent helpers; CharSequence in
  AppLauncherTool; CharSequence/String in FindNodeTool; Network in
  WifiInfoTool; Uuid in PreferencesStore; ListToolsResult in McpManager;
  Assistant in GenerationHandler; TextToSpeech in SystemTTSProvider.
* Drop redundant `!!` non-null assertions on already-non-null String in
  WebDavClient, GoogleProvider, ChatMessageReasoning.
* Drop redundant Elvis `?:` operating on non-null receivers in
  TelegramBotService, AppLauncherTool, SettingNotificationsPage,
  ModelInstall.
* Drop dead `else` branches from exhaustive `when` expressions over sealed
  hierarchies (JsonElement in ErrorParser/JsonExpression, ApprovalScope in
  ChatService).
* Drop redundant `.toString()` / `.toFloat()` conversions where the receiver
  is already the target type (BrowserTools, ViewText).
* Drop redundant `as?` cast where the surrounding flow already proves the
  type (MigrationUtils).

API modernization on safe migrations:
* Rename `MenuAnchorType` to `ExposedDropdownMenuAnchorType` (typealias
  deprecated, drop-in replacement) across TTSProviderConfigure and
  TranslatorPage.
* Migrate `Locale(p0, p1)` and `Locale(p0)` constructors to
  `Locale.forLanguageTag` across ChatMessageTranslation and TranslatorPage.
* Swap `LocalLifecycleOwner` import from
  `androidx.compose.ui.platform` to `androidx.lifecycle.compose` in
  SettingAccessibilityPage and SettingPermissionsPage (same symbol, moved).
* Modernize `getParcelableExtra(name)` to the API 33+ Class overload with a
  legacy fallback in ToolHostActivity and BroadcastTriggers.
* Add `TooltipAnchorPosition.Above` to the new
  `rememberTooltipPositionProvider` signature in Tooltip.
* Swap `kotlin.jvm.Transient` to `kotlinx.serialization.Transient` in
  PreferencesStore's @Serializable Settings class.
* Switch `String.toRequestBody` import: keep only the companion-extension
  form; the bare RequestBody import was unused.
* Promote one-shot `Json { ... }` to a top-level val in SkillCatalog so
  parseSkillCatalogJson doesn't rebuild the configuration per call.
* Block expression body for `TriggerRegistry.resync` plus `return@withLock`
  so the embedded `return` matches the lambda, not the outer function.

Add @OptIn for experimental APIs we genuinely depend on:
* Coil's `setSingletonImageLoaderFactory` (RouteActivity).
* Haze Materials API (file-level on ChatInput).
* `kotlinx.coroutines.FlowPreview.debounce` (ModelList, SearchVM).
* `kotlinx.serialization.ExperimentalSerializationApi.classDiscriminatorMode`
  (McpManager).

Add @Suppress("DEPRECATION") with explanatory comment for APIs that have no
zero-cost replacement at the current dependency versions: ConnectivityManager
allNetworks (SshTool), WebView allowFileAccessFromFileURLs and
allowUniversalAccessFromFileURLs (JsSkillRunner, intentional for sandboxed
JS), WifiInfo connectionInfo/ipAddress (NsdServiceRegistrar, synchronous IP
lookup), ACTION_MEDIA_SCANNER_SCAN_FILE (ContextUtil, gallery refresh path),
MediaSessionCompat flags (MediaPlaybackService), currentWindowDpSize
(ChatPage), LocalClipboardManager (ImgGenPage, SettingWebPage),
FlowRowOverflow (SettingMcpPage file-level), Configuration.locale (AcceptLang),
ByteReadChannel.readUTF8Line (StreamableHttpClientTransport), and the
deprecated UtteranceProgressListener.onError override (SystemTTSProvider).

ai/Message.kt suppresses internally where the exhaustive `when` over
UIMessagePart must list the deprecated ToolCall/ToolResult/Search variants
for type-safety; `ToolCall.merge` does the same for its self-reference.

FastPathRouter's seven `Intent.tryMatch` overrides rename their parameter
from `n` to `normalized` to match the supertype declaration (silences
named-argument warnings).

Build-config cleanups:
* Remove the unresolved `androidx.navigation3.runtime.ExperimentalNavigation3Api`
  opt-in from `app/build.gradle.kts` (the marker class was removed from the
  runtime artifact in a newer navigation3 release; opt-in is no longer
  needed).
* `getByName("androidTest").assets.srcDirs(...)` -> `directories.add(...)`
  per the Gradle DSL deprecation.

Also drops half a dozen unused imports left over from the recent Telegram
refactor: TimeZone/toInstant/booleanOrNull/intOrNull/TelegramAttachment/
add/addJsonArray/addJsonObject/buildJsonArray in TelegramBotService,
RequestBody in TelegramBotClient.

902 JVM tests still green. Release APK installs clean.

+++ app/src/main/java/me/rerere/rikkahub/ui/components/ui/ViewText.kt
@@ -187,7 +187,7 @@ private fun TextView.setComposeTextStyle(
                 // extra = desired_line_height - font_metrics_height
                 val fontMetrics = paint.fontMetricsInt
                 val extraSpacing = lineHeightPx - (fontMetrics.descent - fontMetrics.ascent)
-                setLineSpacing(extraSpacing.toFloat(), 1.0f)
+                setLineSpacing(extraSpacing, 1.0f)
             }
         }
